tests/psoc6: Fix failing test.
