+++ tests/ports/psoc6/hw_ext/sdcard.py
@@ -1,14 +1,19 @@
 import machine
 import os
 import errno
+import vfs
 
+# Define constants
+MOUNT_POINT = "/sd"
 SHORT_TEST_STRING = "This is a test string."
 LONG_TEST_STRING = "This is a very long string. And as a long string that it is, it is only getting longer and longer and the string goes. How long shall it be? Well, not really sure, but let´s try it like this."
+READ_SIZE = 512
+WRITE_SIZE = 512
 
 board = os.uname().machine
 if "CY8CPROTO-062-4343W" in board:
     sdcard_config = {
-        "slot": 1,
+        "slot": 0,
         "width": 4,
         "cd": "P13_5",
         "cmd": "P12_4",
@@ -34,15 +39,22 @@ def unmount_sd_card(path):
             raise Exception(f"Could not unmount {path}")
 
 
-def mount_or_format_sd_card(block_device, filesystem, mount_point, read_size, write_size):
+def mount_or_format_sd_card(block_device, filesystem, mount_point):
     try:
-        vfs = filesystem(block_device, progsize=write_size, readsize=read_size)
+        if filesystem == os.VfsLfs2:
+            vfs = filesystem(block_device, progsize=WRITE_SIZE, readsize=READ_SIZE)
+        else:
+            vfs = filesystem(block_device)
         os.mount(vfs, mount_point)
     except OSError:
-        filesystem.mkfs(block_device, progsize=write_size, readsize=read_size)
-        vfs = filesystem(block_device, progsize=write_size, readsize=read_size)
+        if filesystem == os.VfsLfs2:
+            filesystem.mkfs(block_device, progsize=WRITE_SIZE, readsize=READ_SIZE)
+            vfs = filesystem(block_device, progsize=WRITE_SIZE, readsize=READ_SIZE)
+        else:
+            filesystem.mkfs(block_device)
+            vfs = filesystem(block_device)
         os.mount(vfs, mount_point)
-    print(f"SD card mounted at {mount_point}\n")
+    print(f"\nSD card mounted at {mount_point}\n")
 
 
 def read_write_test(file_path, test_data):
@@ -52,25 +64,19 @@ def read_write_test(file_path, test_data):
         return f.read() == test_data
 
 
-def test_file_transfer():
-    # Define the SD card configuration
+def test_lfs2_file_transfer():
     bdev = machine.SDCard(**sdcard_config)
-
-    # Define constants
-    MOUNT_POINT = "/SDCard"
-    READ_SIZE = 512
-    WRITE_SIZE = 512
 
     # Unmount the SD card if mounted
     unmount_sd_card(MOUNT_POINT)
 
     # Mount or format the SD card with LFS2 filesystem
     if "VfsLfs2" in dir(os):
-        mount_or_format_sd_card(bdev, os.VfsLfs2, MOUNT_POINT, READ_SIZE, WRITE_SIZE)
+        mount_or_format_sd_card(bdev, os.VfsLfs2, MOUNT_POINT)
 
         print("\n***** Test 1: Short string file transfer to SD Card in LFS2 format *****\n")
         # Test short string
-        short_test_file = "/SDCard/test_sd_lfs2_short.txt"
+        short_test_file = MOUNT_POINT + "/test_lfs2_short.txt"
         if read_write_test(short_test_file, SHORT_TEST_STRING):
             print("PASS")
         else:
@@ -78,7 +84,36 @@ def test_file_transfer():
 
         print("\n***** Test 2: Long string file transfer to SD Card in LFS2 format *****\n")
         # Test long string
-        long_test_file = "/SDCard/test_sd_lfs2_long.txt"
+        long_test_file = MOUNT_POINT + "/test_lfs2_long.txt"
+        if read_write_test(long_test_file, LONG_TEST_STRING):
+            print("PASS")
+        else:
+            print("FAIL")
+
+    bdev.deinit()
+
+
+def test_fat_file_transfer():
+    bdev = machine.SDCard(**sdcard_config)
+
+    # Unmount the SD card if mounted
+    unmount_sd_card(MOUNT_POINT)
+
+    # Mount or format the SD card with LFS2 filesystem
+    if "VfsFat" in dir(os):
+        mount_or_format_sd_card(bdev, os.VfsFat, MOUNT_POINT)
+
+        print("\n***** Test 3: Short string file transfer to SD Card in FAT format *****\n")
+        # Test short string
+        short_test_file = MOUNT_POINT + "/test_fat_short.txt"
+        if read_write_test(short_test_file, SHORT_TEST_STRING):
+            print("PASS")
+        else:
+            print("FAIL")
+
+        print("\n***** Test 4: Long string file transfer to SD Card in FAT format *****\n")
+        # Test long string
+        long_test_file = MOUNT_POINT + "/test_fat_long.txt"
         if read_write_test(long_test_file, LONG_TEST_STRING):
             print("PASS")
         else:
@@ -97,5 +132,6 @@ def test_reintializing_same_slot():
 
 
 if __name__ == "__main__":
-    test_file_transfer()
+    # TODO: Run the test based on the enabled file system
+    test_fat_file_transfer()
     test_reintializing_same_slot()
