--- tests/ports/psoc6/sdcard.py.exp
@@ -1,31 +1,0 @@
-
-SD card mounted at /SDCardLFS2
-
-***** Test 1: File transfer to SD Card in LFS2 farmat *****
-
-***** Test 1: Short string file transfer to SD Card in LFS2 format *****
-
-PASS
-
-***** Test 2: Long string file transfer to SD Card in LFS2 format *****
-
-PASS
-
-SD card mounted at /SDCardFAT
-
-
-***** Test 3: Short string file transfer to SD Card in FAT format *****
-
-PASS
-
-***** Test 4: Long string file transfer to SD Card in FAT format *****
-
-PASS
-
-***** Test 5: reinitialize the same slot more than once *****
-
-PASS
-
-***** Test 6: slot number exceeding the number of available slots *****
-
-FAIL
